+++ tests/lib.rs
@@ -1,4 +1,4 @@
 extern crate bson;
 extern crate rustc_serialize;
 
-mod oid;
+mod modules;
