+++ tests/modules/mod.rs
@@ -1,0 +1,3 @@
+mod encoder;
+mod oid;
+mod ordered;
